Rebase and address feedback
